piecrust: adapt runtime to upstream Wasmtime 44
Move the runtime to Wasmtime 44 and point the dependency directly at
upstream `wasmtime` 44.0.2 instead of the `dusk-wasmtime` fork. The v44
LinearMemory API replaces `maximum_byte_size`/`wasm_accessible`, and the
operator-cost configuration moves to `u8` costs.

Drop the fork-only `LinearMemory::needs_init` hook: the one commit earlier added
memory initializer now handles contract-memory init, so the fork becomes obsolete.

+++ piecrust/src/config.rs
@@ -5,4 +5,4 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 // The gas cost for each byte
-pub const BYTE_STORE_COST: i64 = 4;
+pub const BYTE_STORE_COST: u8 = 4;
